extract function resolving logic into dedicated file
https://coveord.atlassian.net/browse/KIT-439
