+++ packages/headless/doc-parser/src/function-param-resolver.ts
@@ -1,5 +1,6 @@
 import {
   ApiEntryPoint,
+  ApiFunction,
   ApiInterface,
   ExcerptTokenKind,
   Parameter,
@@ -9,10 +10,20 @@ import {ObjEntity} from './entity';
 import {buildParamEntity} from './entity-builder';
 import {resolveInterfaceMembers} from './interface-resolver';
 
-export function buildParamEntityBasedOnKind(
-  entryPoint: ApiEntryPoint,
-  p: Parameter
+export function resolveParams(
+  entry: ApiEntryPoint,
+  fn: ApiFunction,
+  shallowParamIndices: number[]
 ) {
+  return fn.parameters.map((p, index) => {
+    const shouldResolveShallow = shallowParamIndices.includes(index);
+    return shouldResolveShallow
+      ? buildParamEntity(p)
+      : buildParamEntityBasedOnKind(entry, p);
+  });
+}
+
+function buildParamEntityBasedOnKind(entryPoint: ApiEntryPoint, p: Parameter) {
   const {kind} = p.parameterTypeExcerpt.spannedTokens[0];
   const isReference = kind === ExcerptTokenKind.Reference;
